watch_service/pyproject.toml
[project]
name = "watch-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "boto3>=1.38.46",
    "fastapi>=0.115.14",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.35.0",
]
